drv:src: Reverting CERT-C changes for Txn Module
This fix reverts the CERT-c changes for txn structures

+++ driver/src/common/xaie_helper.c
@@ -1024,7 +1024,7 @@ static inline void _XAie_AppendBlockWrite32(XAie_DevInst *DevInst,
 	u8 *Payload = TxnPtr + sizeof(XAie_BlockWrite32Hdr);
 	XAie_BlockWrite32Hdr *Hdr = (XAie_BlockWrite32Hdr*)(uintptr_t)TxnPtr;
 
-	Hdr->RegOff = Cmd->RegOff;
+	Hdr->RegOff = (u32)Cmd->RegOff;
 	Hdr->Size = (u32)sizeof(*Hdr) + Cmd->Size * (u32)sizeof(u32);
 	Hdr->OpHdr.Col = _XAie_GetColfromRegOff(DevInst,Cmd->RegOff);
 	Hdr->OpHdr.Row = _XAie_GetRowfromRegOff(DevInst,Cmd->RegOff);
@@ -1045,7 +1045,7 @@ static inline void _XAie_AppendBlockSet32(XAie_DevInst *DevInst,
 	u8 *Payload = TxnPtr + sizeof(XAie_BlockWrite32Hdr);
 	XAie_BlockWrite32Hdr *Hdr = (XAie_BlockWrite32Hdr*)(uintptr_t)TxnPtr;
 
-	Hdr->RegOff = Cmd->RegOff;
+	Hdr->RegOff = (u32)Cmd->RegOff;
 	Hdr->Size = (u32)sizeof(*Hdr) + Cmd->Size * (u32)sizeof(u32);
 	Hdr->OpHdr.Col = _XAie_GetColfromRegOff(DevInst,Cmd->RegOff);
 	Hdr->OpHdr.Row = _XAie_GetRowfromRegOff(DevInst,Cmd->RegOff);
@@ -1148,7 +1148,7 @@ static inline void _XAie_AppendBlockWrite32_opt(XAie_TxnCmd *Cmd, u8 *TxnPtr)
 	u32 *Payload = (void*)(TxnPtr + sizeof(XAie_BlockWrite32Hdr_opt));
 	XAie_BlockWrite32Hdr_opt *Hdr = (XAie_BlockWrite32Hdr_opt*)(uintptr_t)TxnPtr;
 
-	Hdr->RegOff = Cmd->RegOff;
+	Hdr->RegOff = (u32)Cmd->RegOff;
 	Hdr->Size = (u32)sizeof(*Hdr) + Cmd->Size * (u32)sizeof(u32);
 	Hdr->OpHdr.Op = (u8)XAIE_IO_BLOCKWRITE;
 
@@ -1167,7 +1167,7 @@ static inline void _XAie_AppendBlockSet32_opt(XAie_TxnCmd *Cmd, u8 *TxnPtr)
 	u8 *Payload = TxnPtr + sizeof(XAie_BlockWrite32Hdr_opt);
 	XAie_BlockWrite32Hdr_opt *Hdr = (XAie_BlockWrite32Hdr_opt*)(uintptr_t)TxnPtr;
 
-	Hdr->RegOff = Cmd->RegOff;
+	Hdr->RegOff = (u32)Cmd->RegOff;
 	Hdr->Size = (u32)sizeof(*Hdr) + Cmd->Size * (u32)sizeof(u32);
 	Hdr->OpHdr.Op = (u8)XAIE_IO_BLOCKWRITE;
